Move classes from package
