+++ src/main/java/org/cyclonedx/model/component/data/ComponentData.java
@@ -1,4 +1,4 @@
-package org.cyclonedx.model.component.modelCard;
+package org.cyclonedx.model.component.data;
 
 import java.util.List;
 import java.util.Objects;
@@ -8,9 +8,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import org.cyclonedx.model.ExtensibleElement;
-import org.cyclonedx.model.component.modelCard.data.Content;
-import org.cyclonedx.model.component.modelCard.data.Governance;
-import org.cyclonedx.model.component.modelCard.data.Graphics;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
