+++ src/main/java/org/cyclonedx/model/component/data/Governance.java
@@ -1,4 +1,4 @@
-package org.cyclonedx.model.component.modelCard.data;
+package org.cyclonedx.model.component.data;
 
 import java.util.List;
 import java.util.Objects;
