+++ src/main/java/org/cyclonedx/model/component/modelCard/QuantitativeAnalysis.java
@@ -9,7 +9,8 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import org.cyclonedx.model.ExtensibleElement;
-import org.cyclonedx.model.component.modelCard.data.Graphics;
+import org.cyclonedx.model.component.data.Graphics;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 @JsonPropertyOrder({"performanceMetrics", "graphics"})
