+++ src/main/java/org/cyclonedx/util/deserializer/DatasetsChoiceDeserializer.java
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.cyclonedx.model.component.modelCard.ComponentData;
-import org.cyclonedx.model.component.modelCard.DatasetChoice;
+import org.cyclonedx.model.component.data.ComponentData;
+import org.cyclonedx.model.component.data.DatasetChoice;
 
 import java.io.IOException;
